+++ src/main/java/com/phoenix/assetbe/dto/asset/AssetResponse.java
@@ -1,12 +1,9 @@
 package com.phoenix.assetbe.dto.asset;
 
-
 import com.phoenix.assetbe.model.asset.Asset;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public class AssetResponse {
@@ -79,43 +76,6 @@ public class AssetResponse {
             this.visitCount = asset.getVisitCount();
             this.wishlistId = wishlistId;
             this.tagList = tagList;
-        }
-    }
-
-    @Getter @Setter
-    public static class ReviewsOutDTO {
-        private boolean hasAsset;
-        private boolean hasReview;
-        private List<Reviews> reviewList;
-
-        public ReviewsOutDTO(boolean hasAsset, boolean hasReview, List<Reviews> reviewList) {
-            this.hasAsset = hasAsset;
-            this.hasReview = hasReview;
-            this.reviewList = reviewList;
-        }
-
-        @Getter @Setter
-        public static class Reviews {
-            private Long reviewId;
-            private Integer rating;
-            private String content;
-            private Long userId;
-            private String firstName;
-            private String lastName;
-            @JsonIgnore
-            private LocalDateTime createdAt;
-
-            public Reviews(Long reviewId, Integer rating, String content,
-                           Long userId, String firstName, String lastName,
-                           LocalDateTime createdAt) {
-                this.reviewId = reviewId;
-                this.rating = rating;
-                this.content = content;
-                this.userId = userId;
-                this.firstName = firstName;
-                this.lastName = lastName;
-                this.createdAt = createdAt;
-            }
         }
     }
 }
